[#28] refactor: 서비스 로직이 WebClientResponse를 반환하도록 변경
- BotResponse.java의 필요성이 없어짐
