--- src/main/java/com/toy/getyourfriday/service/BotResponse.java
@@ -1,12 +1,0 @@
-package com.toy.getyourfriday.service;
-
-public enum BotResponse {
-    REGISTER_SUCCESS("등록 성공"),
-    REMOVE_SUCCESS("제거 성공");
-
-    private final String message;
-
-    BotResponse(String message) {
-        this.message = message;
-    }
-}
