+++ src/main/java/com/toy/getyourfriday/service/ScrapingService.java
@@ -1,10 +1,11 @@
 package com.toy.getyourfriday.service;
 
+import com.toy.getyourfriday.domain.response.WebClientResponse;
 import com.toy.getyourfriday.dto.RegisterRequest;
 import com.toy.getyourfriday.dto.RemoveRequest;
 
 public interface ScrapingService {
 
-    BotResponse register(RegisterRequest registerRequest);
-    BotResponse remove(RemoveRequest removeRequest);
+    WebClientResponse register(RegisterRequest registerRequest);
+    WebClientResponse remove(RemoveRequest removeRequest);
 }
